modify code according to clippy suggestion

+++ bc/chain/eee/src/extrinsic/xt_primitives.rs
@@ -21,7 +21,7 @@ pub struct GenericExtra(Era, Compact<u32>, Compact<u128>);
 
 impl GenericExtra {
     pub fn new(era: Era, nonce: u32) -> GenericExtra {
-        GenericExtra(era, Compact(nonce), Compact(0 as u128))
+        GenericExtra(era, Compact(nonce), Compact(0_u128))
     }
    pub fn get_nonce(&self)->u32{
         self.1.0
